requirements.txt
absl-py==0.13.0
addict==2.4.0
aenum==3.1.0
alabaster==0.7.12
allure-pytest==2.8.6
allure-python-commons==2.8.6
appdirs==1.4.4
Appium-Python-Client==0.28
argon2-cffi==21.1.0
asgiref==3.2.3
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
basicsr==1.3.3.10
beautifulsoup4==4.9.3
behave==1.2.6
black==19.10b0
bleach==4.1.0
boto3==1.18.27
botocore==1.21.27
cachetools==4.2.2
certifi==2018.4.16
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cheroot==8.5.2
CherryPy==18.6.1
clang==5.0
clear-screen==0.1.14
click==8.0.1
colorama==0.3.9
comtypes==1.1.10
configparser==4.0.2
coverage==4.5.4
coveralls==1.8.2
crayons==0.3.0
crypto==1.4.1
cycler==0.10.0
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
dj-database-url==0.5.0
Django==3.0.3
django-ckeditor==6.1.0
django-classy-tags==2.0.0
django-cms==3.9.0
django-formtools==2.3
django-jet==1.0.8
django-js-asset==1.2.2
django-sekizai==2.0.0
django-summernote==0.8.11.6
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
execnet==1.9.0
extras==1.0.0
facexlib==0.1.1
Faker==8.12.1
fancycompleter==0.9.1
fixtures==3.0.0
flake8==3.7.9
flatbuffers==1.12
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
glob2==0.7
google-api-core==2.0.0
google-api-python-client==2.17.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.19.1
idna==2.7
imageio==2.9.0
imageio-ffmpeg==0.4.5
imagesize==1.2.0
importlib-metadata==0.23
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.4.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.context==4.0.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
jasmine==3.9.0
jasmine-core==3.9.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.1
lemoncheesecake==1.4.0
lettuce==0.2.23
lmdb==1.2.1
loguru==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mediapipe==0.8.7.2
mistune==0.8.4
mock==4.0.3
more-itertools==7.2.0
mox==0.5.3
msedge-selenium-tools==3.141.3
mypy==0.790
mypy-extensions==0.4.3
Naked==0.1.31
natsort==7.1.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.2
node==0.9.26
nose==1.3.7
nose-html-reporting==0.2.3
notebook==6.4.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
odict==1.7.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl==3.0.7
opt-einsum==3.3.0
ordereddict==1.1
packaging==19.2
pandas==1.3.2
pandocfilters==1.4.3
parse==1.19.0
parse-type==0.5.2
parso==0.8.2
pathspec==0.9.0
patool==1.12
pbr==5.6.0
pdbpp==0.10.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.0
plumber==1.6
portend==2.7.1
prometheus-client==0.11.0
prompt-toolkit==2.0.10
protobuf==3.17.3
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.10.1
pyflakes==2.1.1
Pygments==2.10.0
PyHamcrest==2.0.2
pyparsing==2.4.4
PyPDF2==1.26.0
pypiwin32==223
Pypubsub==4.0.3
pyreadline==2.1
pyrepl==0.9.0
pyrsistent==0.18.0
PySimpleGUI==4.46.0
pytesseract==0.3.8
pytest==6.0.1
pytest-clarity==0.2.0a1
pytest-cov==2.8.1
pytest-forked==1.3.0
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-reportportal==5.0.8
pytest-rerunfailures==10.1
pytest-testrail==2.9.0
pytest-xdist==2.3.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.16.0
python-slugify==5.0.2
python-subunit==1.4.0
pyttsx3==2.90
pytz==2019.3
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.4
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
questionary==1.10.0
regex==2021.8.28
reportportal-client==5.0.10
requests==2.26.0
requests-oauthlib==1.3.0
robotframework==4.1
robotframework-ride==2.0b1
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
selenium==3.141.0
self==2020.12.3
Send2Trash==1.8.0
shellescape==3.8.1
six==1.15.0
sklearn==0.0
slackclient==1.2.1
slacker==0.13.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.23
sqlparse==0.3.0
sure==2.0.0
tb-nightly==2.6.0a20210724
tempora==4.1.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
terminaltables==3.1.0
testpath==0.5.0
testtools==2.5.0
tesults==1.2.1
text-unidecode==1.3
threadpoolctl==2.2.0
tifffile==2021.8.8
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.7.4.3
Unipath==1.1
uritemplate==3.0.1
urllib3==1.22
useragent==0.1.1
wcwidth==0.1.7
webdriver-manager==3.4.0
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==2.0.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
win32-setctime==1.0.3
wmctrl==0.4
wolframalpha==5.0.0
wrapt==1.12.1
wxPython==4.1.1
xlrd==1.2.0
xmltodict==0.12.0
yapf==0.31.0
zc.lockfile==2.0
zipp==0.6.0
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
